// src/Polyglot.Frontend/src/app/api/model/chatDto.ts
/**
 * Polyglot.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ChatDto { 
    id?: string;
    title?: string | null;
    createdAt?: string;
    updatedAt?: string;
}